requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
beautifulsoup4==4.14.2
black==25.9.0
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
et-xmlfile==2.0.0
eventlet==0.40.3
faker==37.12.0
filelock==3.20.0
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
fsspec==2025.9.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
keybert==0.9.0
lxml==6.0.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
ollama==0.6.0
openai==2.6.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pypdf2==3.0.1
pypdfium2==5.0.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
rake-nltk==1.0.6
regex==2025.10.23
requests==2.32.5
rich==14.2.0
ruff==0.14.2
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
werkzeug==3.1.3
